+++ jgclark.DashboardReact/requiredFiles/HTMLWinCommsSwitchboard.js
@@ -15,7 +15,7 @@
  */
 // eslint-disable-next-line require-await
 async function delay(time) {
-  return new Promise(resolve => setTimeout(resolve, time))
+  return new Promise((resolve) => setTimeout(resolve, time))
 }
 
 /**
@@ -106,13 +106,13 @@ async function completeTaskInDisplay(data) {
   try {
     const itemID = data.itemID
     console.log(`completeTaskInDisplay: for ID: ${itemID}`)
-    replaceClassInID(`${itemID}I`, "fa-regular fa-circle-check") // adds ticked circle icon
-    addClassToID(itemID, "checked") // adds colour + line-through
-    addClassToID(itemID, "fadeOutAndHide")
+    replaceClassInID(`${itemID}I`, 'fa-regular fa-circle-check') // adds ticked circle icon
+    addClassToID(itemID, 'checked') // adds colour + line-through
+    addClassToID(itemID, 'fadeOutAndHide')
     await delay(1400)
     deleteHTMLItem(itemID)
     // update the totals and other counts
-    incrementItemCount("totalDoneCount")
+    incrementItemCount('totalDoneCount')
     // update the section count(s) if spans with the right ID are present
     const sectionID = itemID.split('-')[0]
     const sectionCountID = `section${sectionID}Count`
@@ -123,7 +123,7 @@ async function completeTaskInDisplay(data) {
     // See if the only remaining item is the '> There are also ... items' line
     const numItemsRemaining = getNumItemsInSectionByClass(`${sectionID}-Section`, 'sectionItemRow')
     if (numItemsRemaining === 1 && doesIDExist(`${sectionID}-Filter`)) {
-    // We need to un-hide the lower-priority items: do full refresh
+      // We need to un-hide the lower-priority items: do full refresh
       console.log(`We need to un-hide the lower-priority items: doing full refresh`)
       sendMessageToPlugin('refresh', { itemID: '', type: '', filename: '', rawContent: '' }) // = actionName, data
     }
@@ -133,7 +133,9 @@ async function completeTaskInDisplay(data) {
       // Delete the whole section from the display
       console.log(`completeTaskInDisplay: trying to delete rest of empty section: ${sectionID}`)
       const sectionItemsGrid = document.getElementById(`${sectionID}-Section`)
-      if (!sectionItemsGrid) { throw new Error(`Couldn't find ID ${itemID}`) }
+      if (!sectionItemsGrid) {
+        throw new Error(`Couldn't find ID ${itemID}`)
+      }
       const enclosingDIV = sectionItemsGrid.parentNode
       console.log(`Will remove node with outerHTML:\n${enclosingDIV.outerHTML}`)
       enclosingDIV.remove()
@@ -151,13 +153,13 @@ async function completeChecklistInDisplay(data) {
   try {
     const itemID = data.itemID
     console.log(`completeChecklistInDisplay: for ID: ${itemID}`)
-    replaceClassInID(`${itemID}I`, "fa-regular fa-square-check") // adds ticked box icon
-    addClassToID(itemID, "checked") // adds colour + line-through text
-    addClassToID(itemID, "fadeOutAndHide")
+    replaceClassInID(`${itemID}I`, 'fa-regular fa-square-check') // adds ticked box icon
+    addClassToID(itemID, 'checked') // adds colour + line-through text
+    addClassToID(itemID, 'fadeOutAndHide')
     await delay(1400)
     deleteHTMLItem(itemID)
     // update the totals
-    incrementItemCount("totalDoneCount")
+    incrementItemCount('totalDoneCount')
     // update the section count(s) if spans with the right ID are present
     const sectionID = itemID.split('-')[0]
     const sectionCountID = `section${sectionID}Count`
@@ -168,7 +170,7 @@ async function completeChecklistInDisplay(data) {
     // See if the only remaining item is the '> There are also ... items' line
     const numItemsRemaining = getNumItemsInSection(`${sectionID}-Section`, 'DIV')
     if (numItemsRemaining === 1 && doesIDExist(`${sectionID}-Filter`)) {
-    // We need to un-hide the lower-priority items: do full refresh
+      // We need to un-hide the lower-priority items: do full refresh
       console.log(`We need to un-hide the lower-priority items: doing full refresh`)
       sendMessageToPlugin('refresh', { itemID: '', type: '', filename: '', rawContent: '' }) // = actionName, data
     }
@@ -178,7 +180,9 @@ async function completeChecklistInDisplay(data) {
       // Delete the whole section from the display
       console.log(`completeChecklistInDisplay: trying to delete rest of empty section: ${sectionID}`)
       const sectionItemsGrid = document.getElementById(`${sectionID}-Section`)
-      if (!sectionItemsGrid) { throw new Error(`Couldn't find ID ${itemID}`) }
+      if (!sectionItemsGrid) {
+        throw new Error(`Couldn't find ID ${itemID}`)
+      }
       const enclosingDIV = sectionItemsGrid.parentNode
       console.log(`Will remove node with outerHTML:\n${enclosingDIV.outerHTML}`)
       enclosingDIV.remove()
@@ -196,9 +200,9 @@ async function cancelTaskInDisplay(data) {
   // const { ID } = data
   const itemID = data.itemID
   console.log(`cancelTaskInDisplay: for ID: ${itemID}`)
-  replaceClassInID(`${itemID}I`, "fa-regular fa-circle-xmark") // adds x-circle icon
-  addClassToID(itemID, "cancelled") // adds colour + line-through text
-  addClassToID(itemID, "fadeOutAndHide")
+  replaceClassInID(`${itemID}I`, 'fa-regular fa-circle-xmark') // adds x-circle icon
+  addClassToID(itemID, 'cancelled') // adds colour + line-through text
+  addClassToID(itemID, 'fadeOutAndHide')
   await delay(1400)
   deleteHTMLItem(itemID)
   // update the section count(s) if spans with the right ID are present
@@ -225,9 +229,9 @@ async function cancelChecklistInDisplay(data) {
   // const { ID } = data
   const itemID = data.itemID
   console.log(`cancelChecklistInDisplay: for ID: ${itemID}`)
-  replaceClassInID(`${itemID}I`, "fa-regular fa-square-xmark") // adds x-box icon
-  addClassToID(itemID, "cancelled") // adds colour + line-through text
-  addClassToID(itemID, "fadeOutAndHide")
+  replaceClassInID(`${itemID}I`, 'fa-regular fa-square-xmark') // adds x-box icon
+  addClassToID(itemID, 'cancelled') // adds colour + line-through text
+  addClassToID(itemID, 'fadeOutAndHide')
   await delay(1400)
   deleteHTMLItem(itemID)
   // update the section count(s) if spans with the right ID are present
@@ -256,12 +260,12 @@ function toggleTypeInDisplay(data) {
   // Get the element with {itemID}I = the icon for that item
   const iconElement = document.getElementById(`${itemID}I`)
   // Switch the icon
-  if (iconElement.className.includes("fa-circle")) {
-    console.log("toggling type to checklist")
-    replaceClassInID(`${itemID}I`, "todo fa-regular fa-square")
+  if (iconElement.className.includes('fa-circle')) {
+    console.log('toggling type to checklist')
+    replaceClassInID(`${itemID}I`, 'todo fa-regular fa-square')
   } else {
-    console.log("toggling type to todo")
-    replaceClassInID(`${itemID}I`, "todo fa-regular fa-circle")
+    console.log('toggling type to todo')
+    replaceClassInID(`${itemID}I`, 'todo fa-regular fa-circle')
   }
 }
 
@@ -321,7 +325,7 @@ function updateItemContent(data) {
 async function unscheduleItem(data) {
   const itemID = data.itemID
   console.log(`unscheduleItem: for ID: ${itemID}`)
-  addClassToID(itemID, "fadeOutAndHide")
+  addClassToID(itemID, 'fadeOutAndHide')
   await delay(1400)
   deleteHTMLItem(itemID)
   // update the section count(s) if spans with the right ID are present
@@ -347,9 +351,7 @@ function setPriorityInDisplay(data) {
   console.log(`- currentInnerHTML: ${currentInnerHTML}`)
 
   // Change the class of the content visible to users, to reflect the new priority colours
-  const newInnerHTML = (data.newPriority > 0)
-    ? `<span class="priority${data.newPriority}">${data.newContent}</span>`
-    : data.newContent
+  const newInnerHTML = data.newPriority > 0 ? `<span class="priority${data.newPriority}">${data.newContent}</span>` : data.newContent
   console.log(`- newInnerHTML: ${newInnerHTML}`)
   replaceHTMLinElement(thisContentElement, newInnerHTML, null)
 
@@ -360,7 +362,7 @@ function setPriorityInDisplay(data) {
   // for (let i = 0; i < tdElements.length; i++) {
   // const tdElement = tdElements[i]
   const tdElement = thisIDElement
-    tdElement.setAttribute('data-encoded-content', data.newContent)
+  tdElement.setAttribute('data-encoded-content', data.newContent)
   console.log(`- set tdElement #${i} data-encoded-content: ${tdElement.getAttribute('data-encoded-content')}`)
   // }
 }
@@ -380,7 +382,7 @@ function setPriorityInDisplay(data) {
  * Note: data is an object
  * @param {Object} data
  */
-function onClickDashboardItem(data) {
+export function onClickDashboardItem(data) {
   sendMessageToPlugin('onClickDashboardItem', data) // actionName, data
 }
 
@@ -475,8 +477,8 @@ function addClassToID(ID, newerClass) {
   // console.log(`addClassToID(${ID}, '${newerClass}') ...`)
   const elem = document.getElementById(ID)
   if (elem) {
-    const origClass = elem.getAttribute("class")
-    elem.setAttribute("class", `${origClass} ${newerClass}`)
+    const origClass = elem.getAttribute('class')
+    elem.setAttribute('class', `${origClass} ${newerClass}`)
   } else {
     console.log(`- ❗error❗ in addClassToID: couldn't find an elem with ID ${ID} to add class ${newerClass}`)
   }
@@ -487,7 +489,7 @@ function replaceClassInID(ID, replacementClass) {
   // console.log(`replaceClassInID(${ID}, '${replacementClass}') ...`)
   const elem = document.getElementById(ID)
   if (elem) {
-    elem.setAttribute("class", replacementClass)
+    elem.setAttribute('class', replacementClass)
   } else {
     console.log(`- error in replaceClassInID: couldn't find an elem with ID ${ID} to replace class ${replacementClass}`)
   }