+++ jgclark.DashboardReact/src/react/components/ItemRow.jsx
@@ -6,6 +6,7 @@
 //--------------------------------------------------------------------------
 import React from 'react'
 import type { TSection, TSectionItem, TParagraphForDashboard } from '../../types.js'
+import { onClickDashboardItem } from '../../../requiredFiles/HTMLWinCommsSwitchboard.js'
 import { useAppContext } from './AppContext.jsx'
 import ItemContent from './ItemContent.jsx'
 import ItemNoteLink from './ItemNoteLink.jsx'
@@ -58,7 +59,8 @@ function ItemRow(inputObj: Props): React$Node {
       const dataObjToPassToOnClick = { itemID: item.ID, type: 'showNoteInEditorFromFilename', encodedFilename: encodedItemFilename, encodedContent: '' }
 
       return (
-        <div className="sectionItemRow"
+        <div
+          className="sectionItemRow"
           id={item.ID}
           data-section-type={sectionType}
           // $FlowIgnore(cannot-resolve-name)
@@ -66,69 +68,64 @@ function ItemRow(inputObj: Props): React$Node {
           // $FlowIgnore(cannot-resolve-name)
           data-encoded-content={encodeRFC3986URIComponent(projectContent)}
         >
-          <div className="reviewProject todo itemIcon"><i id={`${item.ID}I`} className="fa-regular fa-circle-play"></i>
+          <div className="reviewProject todo itemIcon">
+            <i id={`${item.ID}I`} className="fa-regular fa-circle-play"></i>
           </div>
 
           <div className="sectionItemContent sectionItem">
             {/* <a className="content" dangerouslySetInnerHTML={{ __html: projectContent }}></a> */}
             {folderNamePart}
-            <a className="noteTitle sectionItem"
-              onClick={() => onClickDashboardItem(dataObjToPassToOnClick)
-              } >
+            <a className="noteTitle sectionItem" onClick={() => onClickDashboardItem(dataObjToPassToOnClick)}>
               <i className="fa-regular fa-file-lines pad-right"></i>
               {noteTitle}
-            </a >
+            </a>
 
-            <a className="dialogTrigger"
+            <a
+              className="dialogTrigger"
               // $FlowIgnore(cannot-resolve-name)
-              onClick={() => showProjectControlDialog(dataObjectToPassToControlDialog)}>
+              onClick={() => showProjectControlDialog(dataObjectToPassToControlDialog)}
+            >
               <i className="fa-light fa-edit pad-left"></i>
             </a>
           </div>
         </div>
       )
     }
-      // -------------------------------------------------------
+    // -------------------------------------------------------
     else if (itemType === 'congrats') {
       // Display congratulatory message
       return (
-        <div className="sectionItemRow"
-          id={item.ID}
-          data-section-type={sectionType}
-          data-encoded-filename=''
-          data-encoded-content=''
-        >
+        <div className="sectionItemRow" id={item.ID} data-section-type={sectionType} data-encoded-filename="" data-encoded-content="">
           <div className="itemIcon checked">
-            <i id={item.ID}
-              className="fa-regular fa-circle-check"></i>
+            <i id={item.ID} className="fa-regular fa-circle-check"></i>
           </div>
           <div className="sectionItemContent sectionItem">
-            <a className="content"><i>Nothing to do: take a break <i className="fa-regular fa-mug"></i></i></a>
+            <a className="content">
+              <i>
+                Nothing to do: take a break <i className="fa-regular fa-mug"></i>
+              </i>
+            </a>
           </div>
         </div>
       )
     }
-      // -------------------------------------------------------
+    // -------------------------------------------------------
     else if (itemType === 'filterIndicator') {
       // Display filter indicator
       return (
-        <div className="sectionItemRow"
-          id={item.ID}
-          data-section-type={sectionType}
-          data-encoded-filename=''
-          data-encoded-content=''
-        >
+        <div className="sectionItemRow" id={item.ID} data-section-type={sectionType} data-encoded-filename="" data-encoded-content="">
           <div className="itemIcon checked">
-            <i id={item.ID}
-              className="fa-light fa-plus"></i>
+            <i id={item.ID} className="fa-light fa-plus"></i>
           </div>
           <div className="sectionItemContent sectionItem">
-            <a className="content"><i>{item.para.content}</i></a>
+            <a className="content">
+              <i>{item?.para?.content || '<no content>'}</i>
+            </a>
           </div>
         </div>
       )
     }
-      // -------------------------------------------------------
+    // -------------------------------------------------------
     else {
       // Display every other type of item
       if (!item.para) {
@@ -136,20 +133,9 @@ function ItemRow(inputObj: Props): React$Node {
       }
       // compute the things we need later
       // const para: TParagraphForDashboard = item.para
-      const divClassName = (itemType === 'open')
-        ? 'sectionItemTodo todo'
-        : (itemType === 'checklist')
-          ? 'sectionItemChecklist todo'
-          : (itemType === 'congrats')
-            ? 'checked'
-            : ''
-      const iconClassName = (itemType === 'open')
-        ? 'todo fa-regular fa-circle'
-        : (itemType === 'checklist')
-          ? 'todo fa-regular fa-square'
-          : (itemType === 'congrats')
-            ? 'fa-regular fa-circle-check'
-            : ''
+      const divClassName = itemType === 'open' ? 'sectionItemTodo todo' : itemType === 'checklist' ? 'sectionItemChecklist todo' : itemType === 'congrats' ? 'checked' : ''
+      const iconClassName =
+        itemType === 'open' ? 'todo fa-regular fa-circle' : itemType === 'checklist' ? 'todo fa-regular fa-square' : itemType === 'congrats' ? 'fa-regular fa-circle-check' : ''
       const dataObjectToPassToControlDialog = {
         OS: 'macOS', // TODO: NotePlan.environment.platform,
         itemID: item.ID,
@@ -160,7 +146,8 @@ function ItemRow(inputObj: Props): React$Node {
       }
 
       return (
-        <div className="sectionItemRow"
+        <div
+          className="sectionItemRow"
           id={item.ID}
           data-section-type={sectionType}
           // $FlowIgnore(cannot-resolve-name)
@@ -168,15 +155,18 @@ function ItemRow(inputObj: Props): React$Node {
           // $FlowIgnore(cannot-resolve-name)
           data-encoded-content={encodeRFC3986URIComponent(item.para?.content)}
         >
-          <div className={`${divClassName} itemIcon`}><i id={`${item.ID}I`} className={`${iconClassName}`}></i>
+          <div className={`${divClassName} itemIcon`}>
+            <i id={`${item.ID}I`} className={`${iconClassName}`}></i>
           </div>
 
           <div className="sectionItemContent sectionItem">
             <ItemContent item={item} />
-            {(config.includeTaskContext) ? <ItemNoteLink item={item} thisSection={thisSection} /> : null}
-            <a className="dialogTrigger"
+            {config.includeTaskContext ? <ItemNoteLink item={item} thisSection={thisSection} /> : null}
+            <a
+              className="dialogTrigger"
               // $FlowIgnore(cannot-resolve-name)
-              onClick={() => showItemControlDialog(dataObjectToPassToControlDialog)}>
+              onClick={() => showItemControlDialog(dataObjectToPassToControlDialog)}
+            >
               <i className="fa-light fa-edit pad-left"></i>
             </a>
           </div>
@@ -199,7 +189,9 @@ function makeNoteTitleWithOpenActionFromFilename(item: TSectionItem, noteTitle: 
   try {
     // console.log(`makeNoteTitleWithOpenActionFromFilename: - making notelink with ${item.filename}, ${noteTitle}`)
     // Pass request back to plugin, as a single object
-    return `<a class="noteTitle sectionItem" onClick="onClickDashboardItem({itemID: '${item.ID}', type: 'showNoteInEditorFromFilename', encodedFilename: '${encodeURIComponent(item.itemFilename)}', encodedContent: ''})"><i class="fa-regular fa-file-lines pad-right"></i> ${noteTitle}</a>`
+    return `<a class="noteTitle sectionItem" onClick="onClickDashboardItem({itemID: '${item.ID}', type: 'showNoteInEditorFromFilename', encodedFilename: '${encodeURIComponent(
+      item.itemFilename,
+    )}', encodedContent: ''})"><i class="fa-regular fa-file-lines pad-right"></i> ${noteTitle}</a>`
   } catch (error) {
     console.error(`❗️ERROR❗️ ItemRow::makeNoteTitleWithOpenActionFromFilename: ${error.message} for input '${noteTitle}'`)
     return '(error)'