+++ jgclark.DashboardReact/src/reactMain.js
@@ -7,11 +7,8 @@
 // import moment from 'moment/min/moment-with-locales'
 import moment from 'moment/min/moment-with-locales'
 import pluginJson from '../plugin.json'
-import {
-  getSettings,
-  type dashboardConfigType,
-} from './dashboardHelpers'
-// import { bridgeClickDashboardItem, bridgeChangeCheckbox, runPluginCommand } from './pluginToHTMLBridge'
+import { getSettings, type dashboardConfigType } from './dashboardHelpers'
+import { bridgeClickDashboardItem, bridgeChangeCheckbox, runPluginCommand } from './pluginToHTMLBridge'
 import type { TSection } from './types'
 import {
   getTodaySectionData,
@@ -108,22 +105,18 @@ export async function showDashboardReact(callMode: string = 'full', demoMode: bo
       windowTitle: data.title,
       customId: WEBVIEW_WINDOW_ID,
       makeModal: false,
-      savedFilename: `../../${pluginJson['plugin.id']}/dashboard-react.html`, /* for saving a debug version of the html file */
-      shouldFocus: callMode !== 'refresh', /* focus window every time (unless this is a refresh) */
+      savedFilename: `../../${pluginJson['plugin.id']}/dashboard-react.html` /* for saving a debug version of the html file */,
+      shouldFocus: callMode !== 'refresh' /* focus window every time (unless this is a refresh) */,
       headerTags: `${resourceLinksInHeader}\n<meta name="startTime" content="${String(Date.now())}">`,
       generalCSSIn: generateCSSFromTheme(config.dashboardTheme), // either use dashboard-specific theme name, or get general CSS set automatically from current theme
       specificCSS: '', // set in separate CSS file referenced in header
-      preBodyScript: `
-      <script type="text/javascript" src="./showTimeAgo.js"></script>
-`, // some extra pre-JS
-      bodyOptions: 'onload="showTimeAgo()"',
-      // postBodyScript: `${commsBridge}
+      preBodyScript: ``,
       postBodyScript: `
       <script type="text/javascript" src="../np.Shared/encodeDecode.js"></script>
       <script type="text/javascript" src="../np.Shared/shortcut.js"></script>
       <script type="text/javascript" src="./dashboardShortcuts.js"></script>
       <script type="text/javascript" src="./dashboardEvents.js"></script>
-`
+`,
     }
     logDebug(`===== showDashboardReact Calling React after ${timer(data.startTime || new Date())} =====`)
     // clo(data, `showDashboardReact data object passed`)
@@ -156,8 +149,7 @@ export async function getInitialDataForReactWindowObjectForReactView(useDemoData
       startTime,
     }
     return dataToPass
-  }
-  catch (error) {
+  } catch (error) {
     logError(pluginJson, error.message)
     return
   }
@@ -185,9 +177,9 @@ export async function getInitialDataForReactWindow(config: dashboardConfigType, 
   }
 }
 
-/** 
- * Get all the sections' data (that the user wants) 
-*/
+/**
+ * Get all the sections' data (that the user wants)
+ */
 async function getAllSectionsData(config: dashboardConfigType, demoMode: boolean = false): Promise<Array<TSection>> {
   const data: Array<TSection> = []
   data.push(getTodaySectionData(config, demoMode))
@@ -234,25 +226,29 @@ export async function updateReactWindowData(actionType: string, data: any = null
 export async function onMessageFromHTMLView(actionType: string, data: any): Promise<any> {
   try {
     logDebug(pluginJson, `NP Plugin return path (onMessageFromHTMLView) received actionType="${actionType}" (typeof=${typeof actionType})  (typeof data=${typeof data})`)
-    clo(data, `Plugin onMessageFromHTMLView data=`)
+    // clo(data, `Plugin onMessageFromHTMLView data=`)
     let reactWindowData = await getGlobalSharedData(WEBVIEW_WINDOW_ID) // get the current data from the React Window
-    clo(reactWindowData, `Plugin onMessageFromHTMLView reactWindowData=`)
     if (data.passThroughVars) reactWindowData.passThroughVars = { ...reactWindowData.passThroughVars, ...data.passThroughVars }
     switch (actionType) {
       /* best practice here is not to actually do the processing but to call a function based on what the actionType was sent by React */
-      case 'refresh':
-        await showDashboardReact('refresh')
+      case 'refresh': {
+        const data = await getInitialDataForReactWindowObjectForReactView(demoMode)
+        clo(reactWindowData, `\n\n\nPlugin onMessageFromHTMLView reactWindowData=`)
+        clo(data, `\n\n\nPlugin onMessageFromHTMLView data=`)
+        reactWindowData = { ...reactWindowData, ...{ pluginData: data } }
+        throw 'stoppping'
         break
-      case 'onSubmitClick':
-        reactWindowData = await handleSubmitButtonClick(data, reactWindowData) //update the data to send it back to the React Window
-        break
+      }
+      // NOTE: SO THAT JGCLARK DOESN'T HAVE TO RE-INVENT THE WHEEL HERE, WE WILL JUST CALL THE PRE-EXISTING FUNCTION bridgeDashboardItem
+      // every time
       default:
-        await sendBannerMessage(WEBVIEW_WINDOW_ID, `Plugin received an unknown actionType: "${actionType}" command with data:\n${JSON.stringify(data)}`)
+        reactWindowData = (await bridgeClickDashboardItem(data)) || reactWindowData // the processing function can update the reactWindowData object and return it
+        // await sendBannerMessage(WEBVIEW_WINDOW_ID, `Plugin received an unknown actionType: "${actionType}" command with data:\n${JSON.stringify(data)}`)
         break
     }
     if (reactWindowData) {
       const updateText = `After ${actionType}, data was updated` /* this is just a string for debugging so you know what changed in the React Window */
-      clo(reactWindowData, `Plugin onMessageFromHTMLView after updating window data,reactWindowData=`)
+      // clo(reactWindowData, `Plugin onMessageFromHTMLView after updating window data,reactWindowData=`)
       sendToHTMLWindow(WEBVIEW_WINDOW_ID, 'SET_DATA', reactWindowData, updateText) // note this will cause the React Window to re-render with the currentJSData
     }
     return {} // this return value is ignored but needs to exist or we get an error
@@ -260,7 +256,6 @@ export async function onMessageFromHTMLView(actionType: string, data: any): Prom
     logError(pluginJson, JSP(error))
   }
 }
-
 
 /**
  * An example handler function that is called when someone clicks a button in the React Window
@@ -285,4 +280,3 @@ async function handleSubmitButtonClick(data: any, reactWindowData: PassedData): 
   reactWindowData.pluginData.tableRows[index].textValue = `Item ${clickedIndex} was updated by the plugin (see changed data in the debug section below)`
   return reactWindowData //updated data to send back to React Window
 }
-